split up and reorganisd hdk_graph_helpers better, pulled RPC-related methods into own module. No code changes.
