+++ lib/hdk_graph_helpers/src/lib.rs
@@ -11,12 +11,14 @@ mod maybe_undefined;
 pub mod record_helpers;
 pub mod record_interface;
 pub mod link_helpers;
+pub mod rpc_helpers;
 
 // submodule exports
 
 pub use maybe_undefined::MaybeUndefined as MaybeUndefined;
 pub use record_helpers as records;
 pub use link_helpers as links;
+pub use rpc_helpers as rpc;
 
 pub mod identifiers {
     pub const RECORD_INITIAL_ENTRY_LINK_TAG: &str = "initial_entry";
